added express module, moved jsx files to component folder
